[refactor] confirm write image slide input component 분리
